+++ native/src/main/java/io/ballerina/stdlib/websocket/WebSocketConstants.java
@@ -60,7 +60,9 @@ public class WebSocketConstants {
     public static final String RESOURCE_NAME_ON_BINARY_MESSAGE = "onBinaryMessage";
     public static final String RESOURCE_NAME_ON_MESSAGE = "onMessage";
     public static final String RESOURCE_NAME_ON_PING = "onPing";
+    public static final String RESOURCE_NAME_ON_PING_MESSAGE = "onPingMessage";
     public static final String RESOURCE_NAME_ON_PONG = "onPong";
+    public static final String RESOURCE_NAME_ON_PONG_MESSAGE = "onPongMessage";
     public static final String RESOURCE_NAME_ON_CLOSE = "onClose";
     public static final String RESOURCE_NAME_ON_IDLE_TIMEOUT = "onIdleTimeout";
     public static final String RESOURCE_NAME_ON_ERROR = "onError";
@@ -139,6 +141,14 @@ public class WebSocketConstants {
     public static final String CONNECTION_CLOSED = "Connection closed";
     public static final String STATUS_CODE = "Status code:";
 
+    // Close Frame Records
+    public static final BString CLOSE_FRAME_TYPE = StringUtils.fromString("type");
+    public static final BString CLOSE_FRAME_STATUS_CODE = StringUtils.fromString("status");
+    public static final BString CLOSE_FRAME_REASON = StringUtils.fromString("reason");
+    public static final String PREDEFINED_CLOSE_FRAME_TYPE = "PredefinedCloseFrameType";
+    public static final String CUSTOM_CLOSE_FRAME_TYPE = "CustomCloseFrameType";
+    public static final int CLOSE_FRAME_DEFAULT_TIMEOUT = 60;
+
     private WebSocketConstants() {
     }
 
@@ -150,7 +160,7 @@ public class WebSocketConstants {
         ConnectionClosureError("ConnectionClosureError"),
         InvalidHandshakeError("InvalidHandshakeError"),
         PayloadTooLargeError("PayloadTooLargeError"),
-        ProtocolError("ProtocolError"),
+        CorruptedFrameError("CorruptedFrameError"),
         ConnectionError("ConnectionError"),
         InvalidContinuationFrameError("InvalidContinuationFrameError"),
         HandshakeTimedOut("HandshakeTimedOut"),